디버그 에러 수동 수정

+++ api/src/main/java/com/vibe/pay/domain/pointhistory/dto/PointHistoryResponse.java
@@ -41,13 +41,13 @@ public class PointHistoryResponse {
      * 포인트 변동 금액 (양수: 적립, 음수: 사용)
      */
     @Schema(description = "포인트 변동 금액")
-    private Integer pointAmount;
+    private Long pointAmount;
 
     /**
      * 변동 후 잔액
      */
     @Schema(description = "변동 후 잔액")
-    private Integer balanceAfter;
+    private Long balanceAfter;
 
     /**
      * 거래 유형 (EARN, USE, REFUND 등)
